Adding 'correct' license file
Back in 2013 or so, I uploaded the wrong license by mistake. On the Google Code Archive, it's clear that the intended license was the Artistic License (it even has a proper link to it). But I believe I might just have seen _AL/GPL_ and wrongly assumed that this was the Affero GPL. My apologies to Pleiades.