Chore: fix PHP to run under modern versions
The many PHP scripts included to test the API have been designed for PHP 4 (with some slight retouching for PHP 5) and now are really requiring an overhaul to keep them up-to-date with current PHP versions and modern idiomatic practices.
Also... they need more documentation and un update on the API to conform with the latest changes!